% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate Body Mass Index (BMI)}
\usage{
calculate_bmi(weight, height)
}
\arguments{
\item{weight}{Weight in kilograms.}

\item{height}{Height in centimeters.}
}
\value{
BMI value.
}
\description{
This function calculates the BMI given weight and height.
}
