% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_sodium_sieving}
\alias{calculate_sodium_sieving}
\title{Calculate Sodium Sieving}
\usage{
calculate_sodium_sieving(sodium0h, sodium1h)
}
\arguments{
\item{sodium0h}{Sodium at 0 hours (mEq/L).}

\item{sodium1h}{Sodium at 1 hour (mEq/L).}
}
\value{
Sodium sieving.
}
\description{
This function calculates sodium sieving.
}
