% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_uf}
\alias{calculate_uf}
\title{Calculate Ultrafiltration (UF) Volume}
\usage{
calculate_uf(drain_volume, infusion_volume)
}
\arguments{
\item{drain_volume}{Drained volume in mL.}

\item{infusion_volume}{Infused volume in mL.}
}
\value{
UF volume.
}
\description{
This function calculates the UF volume given drain and infusion volumes.
}
