% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generatePlotWithBands2}
\alias{generatePlotWithBands2}
\title{Generate Glucose Plot with Bands}
\usage{
generatePlotWithBands2(data, title, yAxisLabel, bandsData)
}
\arguments{
\item{data}{Data frame with time and normalized values.}

\item{title}{Plot title.}

\item{yAxisLabel}{Y-axis label.}

\item{bandsData}{Data for the bands.}
}
\value{
ggplot object.
}
\description{
This function generates a glucose plot with bands.
}
