% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runPeritonealEquilibriumTestApp}
\alias{runPeritonealEquilibriumTestApp}
\title{Run the Peritoneal Equilibrium Test Shiny App}
\usage{
runPeritonealEquilibriumTestApp()
}
\description{
This function launches the Shiny app for the Peritoneal Equilibrium Test.
}
